Extract CollapsingToolbar Code into CollapsingToolbar Composable
