+++ app/src/main/java/de/schnettler/scrobbler/MainActivity.kt
@@ -133,6 +133,7 @@ class MainActivity : AppCompatActivity() {
             is UIAction.ListingSelected -> onListingClicked(action.listing)
             is UIAction.TagSelected -> onTagClicked(action.id)
             is UIAction.TrackLiked -> detailsViewModel.onToggleLoveTrackClicked(action.track, action.info)
+            is UIAction.NavigateUp -> backStackController.pop()
         }
     }
